.cursor/skills/image-bg-remove-local/requirements.txt
mcp
rembg
Pillow
